% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_liu.R
\name{theme_liu}
\alias{theme_liu}
\title{Title This is the theme for ggplot2 using Linkoping University colors}
\usage{
theme_liu(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base parameters}

\item{base_family}{base parameters}
}
\value{
None
}
\description{
Title This is the theme for ggplot2 using Linkoping University colors
}
